Rename WindowConfig
