+++ src/shell/js/composeui-node-launcher/src/index.ts
@@ -1,5 +1,5 @@
 ﻿export { BrowserWindow } from './BrowserWindow.js';
 export { Launcher } from './launcher.js';
-export { WindowConfig } from './WindowConfig.js';
+export { WindowConfig } from './windowConfig.js';
 export * from './cli/cli.js';
 export { executeScriptFile } from './cli/executeScrtiptFile.js';
