+++ src/shell/js/composeui-node-launcher/src/launcher.ts
@@ -1,5 +1,5 @@
 import { execFile } from 'child_process';
-import { WindowConfig } from './WindowConfig';
+import { WindowConfig } from './windowConfig';
 
 export class Launcher {
     private processArgs(config?: WindowConfig) {
