+++ src/shell/js/composeui-node-launcher/src/windowConfig.ts
